backend/requirements.txt
fastapi[all]
langchain
langchain-chroma==0.1.4
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchain-experimental
pypdfium2
tiktoken
open-clip-torch 
tqdm
pdfminer-six
python-docx
loguru
chromadb==0.5.23
Unidecode
langserve[all]
pillow
torch
whisper
pyttsx3
pyaudio
numpy==1.26.0
tokenizers==0.20.3
grpcio==1.68.1
grpcio-status==1.68.1
grpcio-tools==1.68.1
sentence-transformers